+++ src/pages/home/index.js
@@ -1,7 +1,7 @@
 import { useTranslation } from 'react-i18next'
 import { faChevronCircleRight } from '@fortawesome/free-solid-svg-icons'
 import { HomePageContainer, HomeContainer } from './styled'
-import { Links } from '../../components/Links'
+import { Hyperlink } from '../../components/hyperlink'
 
 export const HomePage = () => {
   const { t } = useTranslation()
@@ -9,8 +9,8 @@ export const HomePage = () => {
     <HomePageContainer>
       <h1>{t('home.title')}</h1>
       <HomeContainer>
-        <Links link="/hiring-process" label={t('hiringProcess.title')} icon={faChevronCircleRight} />
-        <Links link="/user" label={t('home.manageMentor')} icon={faChevronCircleRight} />
+        <Hyperlink link="/hiring-process" label={t('hiringProcess.title')} icon={faChevronCircleRight} />
+        <Hyperlink link="/user" label={t('home.manageMentor')} icon={faChevronCircleRight} />
       </HomeContainer>
     </HomePageContainer>
   )
